Added a way to set and retrieve animation curves to AnimationClip in C#
Former-commit-id: b5b2b76c832368937dcdec6d17d9047890c035b3

+++ Source/BansheeEditor/Source/BsEditorUtility.cpp
@@ -87,7 +87,7 @@ namespace BansheeEngine
 		}
 
 		if (gotOneMesh)
-			return center / count;
+			return center / (float)count;
 
 		return Vector3::ZERO;
 	}
